requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
qrcode[pil]>=7.4.2
matplotlib>=3.7.0
fpdf2>=2.7.4
reportlab>=4.0.0